oh-my-guard/requirements.txt
# Oh-My-Guard! Server – Python 3.12 Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
jinja2==3.1.4
aiofiles==24.1.0
websockets==14.1
httpx==0.28.1

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# ── Cache / Real-time ─────────────────────────────────────
redis==5.2.1
aioredis==2.0.1

# ── Cryptography ──────────────────────────────────────────
cryptography==44.0.0
pyopenssl==24.3.0

# ── File Monitoring ───────────────────────────────────────
watchdog==6.0.0
psutil==6.1.0

# ── Packet Inspection (IDS/IPS) ───────────────────────────
scapy==2.6.1
# netfilterqueue==1.1.0  # Linux only – installed by install.sh if on Linux

# ── Auth & Sessions ───────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Email Notifications ───────────────────────────────────
aiosmtplib==3.0.2

# ── Background Tasks ──────────────────────────────────────
celery==5.4.0

# ── Data Validation ───────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
rich==13.9.4
typer==0.15.1
loguru==0.7.3

# ── Testing ───────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1

# ── Security Auditing ─────────────────────────────────────
bandit==1.8.0
safety==3.2.14

# ── Code Quality ──────────────────────────────────────────
mypy==1.14.0
black==24.12.0
isort==5.13.2
